+++ naslib/defaults/trainer_multifidelity.py
@@ -16,7 +16,7 @@ from naslib.utils import utils
 from naslib.utils.logging import log_every_n_seconds, log_first_n
 
 from typing import Callable
-from .additional_primitives import DropPathWrapper #Can be causes issues
+#from .additional_primitives import DropPathWrapper #Can be causes issues
 
 logger = logging.getLogger(__name__)
