+++ naslib/predictors/multifidelity/__init__.py
@@ -1,0 +1,1 @@
+from .multifidelity import TreeParserEstimator
